+++ FVM/IsotropicNeohookeanMaterial.h
@@ -11,22 +11,22 @@ public:
 	IsotropicNeohookeanMaterial(std::shared_ptr<TetMesh> tetMesh);
 	virtual ~IsotropicNeohookeanMaterial();
 
-	virtual Eigen::Vector3d computeEnergy2InvariantsGradient(int tetID, Eigen::Vector3d invariants);
-	virtual Eigen::Matrix3d computeEnergy2InvariantsHessian(int tetID, Eigen::Vector3d invariants);
+	virtual Eigen::Vector3f computeEnergy2InvariantsGradient(int tetID, Eigen::Vector3f invariants);
+	virtual Eigen::Matrix3f computeEnergy2InvariantsHessian(int tetID, Eigen::Vector3f invariants);
 
 protected:
-	virtual double fEnergy(double x, const double & mu, const double & lambda) { return 0.5*mu *(x*x - 1); }
-	virtual double gEnergy(double x, const double & mu, const double & lambda) { return 0; }
-	virtual double hEnergy(double x, const double & mu, const double & lambda)
+	virtual float fEnergy(float x, const float & mu, const float & lambda) { return 0.5*mu *(x*x - 1); }
+	virtual float gEnergy(float x, const float & mu, const float & lambda) { return 0; }
+	virtual float hEnergy(float x, const float & mu, const float & lambda)
 	{
-		double lgx = std::log(x);
+		float lgx = std::log(x);
 		return 0.5 * lambda * (lgx*lgx) - mu*lgx;
 	}
-	virtual double dfEnergy(double x, const double & mu, const double & lambda) { return mu * x; }
-	virtual double dgEnergy(double x, const double & mu, const double & lambda) { return 0; }
-	virtual double dhEnergy(double x, const double & mu, const double & lambda) { return (lambda*std::log(x) - mu) / x; }
-	virtual double ddfEnergy(double x, const double & mu, const double & lambda) { return mu; }
-	virtual double ddgEnergy(double x, const double & mu, const double & lambda) { return 0; }
-	virtual double ddhEnergy(double x, const double & mu, const double & lambda) { return (mu + lambda - lambda*std::log(x)) / (x*x); }
+	virtual float dfEnergy(float x, const float & mu, const float & lambda) { return mu * x; }
+	virtual float dgEnergy(float x, const float & mu, const float & lambda) { return 0; }
+	virtual float dhEnergy(float x, const float & mu, const float & lambda) { return (lambda*std::log(x) - mu) / x; }
+	virtual float ddfEnergy(float x, const float & mu, const float & lambda) { return mu; }
+	virtual float ddgEnergy(float x, const float & mu, const float & lambda) { return 0; }
+	virtual float ddhEnergy(float x, const float & mu, const float & lambda) { return (mu + lambda - lambda*std::log(x)) / (x*x); }
 };
