+++ FVM/TimeIntegration.cpp
@@ -5,17 +5,17 @@
 
 TimeIntegration::TimeIntegration(int num_nodes) : m_t(1e-6)
 {
-	m_positions = Eigen::MatrixXd::Zero(3, num_nodes);
-	m_velocities = Eigen::MatrixXd::Zero(3, num_nodes);
+	m_positions = Eigen::MatrixXf::Zero(3, num_nodes);
+	m_velocities = Eigen::MatrixXf::Zero(3, num_nodes);
 	n_nodes = num_nodes;
 }
 
-TimeIntegration::TimeIntegration(int num_nodes, Eigen::VectorXd m) : m_t(2e-3)
+TimeIntegration::TimeIntegration(int num_nodes, Eigen::VectorXf m) : m_t(2e-3)
 {
-	m_positions = Eigen::MatrixXd::Zero(3, num_nodes);
-	m_velocities = Eigen::MatrixXd::Zero(3, num_nodes);
+	m_positions = Eigen::MatrixXf::Zero(3, num_nodes);
+	m_velocities = Eigen::MatrixXf::Zero(3, num_nodes);
 	n_nodes = num_nodes;
-	m_masses = Eigen::VectorXd::Zero(3 * num_nodes);
+	m_masses = Eigen::VectorXf::Zero(3 * num_nodes);
 
 	for (size_t i = 0; i < num_nodes; ++i)
 	{
@@ -26,13 +26,13 @@ TimeIntegration::TimeIntegration(int num_nodes, Eigen::VectorXd m) : m_t(2e-3)
 
 }
 
-TimeIntegration::TimeIntegration(int num_nodes, Eigen::VectorXd m, std::vector<int> constraints, Eigen::MatrixXd rest_pos) :
+TimeIntegration::TimeIntegration(int num_nodes, Eigen::VectorXf m, std::vector<int> constraints, Eigen::MatrixXf rest_pos) :
 m_t(3e-2), m_constraints(constraints), m_rest(rest_pos), m_dumpingAlpha(0.02), m_dumpingBelta(0.02)
 {
-	m_positions = Eigen::MatrixXd::Zero(3, num_nodes);
-	m_velocities = Eigen::MatrixXd::Zero(3, num_nodes);
+	m_positions = Eigen::MatrixXf::Zero(3, num_nodes);
+	m_velocities = Eigen::MatrixXf::Zero(3, num_nodes);
 	n_nodes = num_nodes;
-	m_masses = Eigen::VectorXd::Zero(3 * num_nodes);
+	m_masses = Eigen::VectorXf::Zero(3 * num_nodes);
 
 	for (size_t i = 0; i < num_nodes; ++i)
 	{
@@ -54,10 +54,10 @@ TimeIntegration::~TimeIntegration()
 {
 }
 
-void TimeIntegration::simuExplicit(	const Eigen::MatrixXd & pos,
-									const Eigen::MatrixXd & vel,
-									const Eigen::MatrixXd & force,
-									const Eigen::VectorXd & mass)
+void TimeIntegration::simuExplicit(	const Eigen::MatrixXf & pos,
+									const Eigen::MatrixXf & vel,
+									const Eigen::MatrixXf & force,
+									const Eigen::VectorXf & mass)
 {
 	for (int i = 0; i < n_nodes; ++i)
 	{
@@ -66,25 +66,25 @@ void TimeIntegration::simuExplicit(	const Eigen::MatrixXd & pos,
 	}
 }
 
-void TimeIntegration::BackEuler( Eigen::MatrixXd & pos,
-								 Eigen::MatrixXd & vel,
-								 Eigen::MatrixXd & force,
-								 Eigen::MatrixXd & K)
+void TimeIntegration::BackEuler( Eigen::MatrixXf & pos,
+								 Eigen::MatrixXf & vel,
+								 Eigen::MatrixXf & force,
+								 Eigen::MatrixXf & K)
 {
 	int n = pos.cols();
-	Eigen::MatrixXd A = m_t*m_t*K;
+	Eigen::MatrixXf A = m_t*m_t*K;
 	A.diagonal() += m_masses;
 	//std::cout << A << std::endl;
 	pos.resize(3*n,1);
 	vel.resize(3*n,1);
 	force.resize(3*n,1);
-	Eigen::VectorXd p = pos;
-	Eigen::VectorXd v = vel;
-	Eigen::VectorXd f = force;
+	Eigen::VectorXf p = pos;
+	Eigen::VectorXf v = vel;
+	Eigen::VectorXf f = force;
 
-	Eigen::VectorXd b = m_masses.cwiseProduct(v) + m_t * f;
+	Eigen::VectorXf b = m_masses.cwiseProduct(v) + m_t * f;
 
-	Eigen::ConjugateGradient<Eigen::MatrixXd> cg_solver;
+	Eigen::ConjugateGradient<Eigen::MatrixXf> cg_solver;
 	cg_solver.compute(A);
 	cg_solver.setTolerance(1e-8);
 	v = cg_solver.solve(b);
@@ -99,20 +99,20 @@ void TimeIntegration::BackEuler( Eigen::MatrixXd & pos,
 	addGroundConstraints(-2.0, pos, vel);
 }
 
-void TimeIntegration::BackEuler(Eigen::MatrixXd & pos,
-	Eigen::MatrixXd & vel,
-	Eigen::MatrixXd & force,
-	Eigen::SparseMatrix<double> & K)
+void TimeIntegration::BackEuler(Eigen::MatrixXf & pos,
+	Eigen::MatrixXf & vel,
+	Eigen::MatrixXf & force,
+	Eigen::SparseMatrix<float> & K)
 {
 	for (int i = 0; i < m_constraints.size(); ++i)
 	{
-		pos.col(m_constraints[i]) = Eigen::Vector3d::Zero();
-		vel.col(m_constraints[i]) = Eigen::Vector3d::Zero();
-		force.col(m_constraints[i]) = Eigen::Vector3d::Zero();
+		pos.col(m_constraints[i]) = Eigen::Vector3f::Zero();
+		vel.col(m_constraints[i]) = Eigen::Vector3f::Zero();
+		force.col(m_constraints[i]) = Eigen::Vector3f::Zero();
 	}
 
 	int n = pos.cols();
-	Eigen::SparseMatrix<double> A = (m_dumpingBelta+m_t)*m_t*K;
+	Eigen::SparseMatrix<float> A = (m_dumpingBelta+m_t)*m_t*K;
 	for (int i = 0; i < 3 * n; ++i)
 	{
 		A.coeffRef(i, i) += (1+m_t*m_dumpingAlpha)*m_masses(i);
@@ -121,13 +121,13 @@ void TimeIntegration::BackEuler(Eigen::MatrixXd & pos,
 	pos.resize(3 * n, 1);
 	vel.resize(3 * n, 1);
 	force.resize(3 * n, 1);
-	Eigen::VectorXd p = pos;
-	Eigen::VectorXd v = vel;
-	Eigen::VectorXd f = force;
+	Eigen::VectorXf p = pos;
+	Eigen::VectorXf v = vel;
+	Eigen::VectorXf f = force;
 
-	Eigen::VectorXd b = m_masses.cwiseProduct(v) + m_t * f;
+	Eigen::VectorXf b = m_masses.cwiseProduct(v) + m_t * f;
 
-	Eigen::ConjugateGradient<Eigen::SparseMatrix<double>> cg_solver;
+	Eigen::ConjugateGradient<Eigen::SparseMatrix<float>> cg_solver;
 	cg_solver.compute(A);
 	cg_solver.setTolerance(1e-8);
 	v = cg_solver.solve(b);
@@ -146,32 +146,32 @@ void TimeIntegration::BackEuler(Eigen::MatrixXd & pos,
 	//addGroundConstraints(-1.0, pos, vel);
 }
 
-void TimeIntegration::BackEuler(Eigen::MatrixXd & pos,
-	Eigen::MatrixXd & restPos,
-	Eigen::MatrixXd & vel,
-	Eigen::MatrixXd & force,
-	Eigen::SparseMatrix<double> & K)
+void TimeIntegration::BackEuler(Eigen::MatrixXf & pos,
+	Eigen::MatrixXf & restPos,
+	Eigen::MatrixXf & vel,
+	Eigen::MatrixXf & force,
+	Eigen::SparseMatrix<float> & K)
 {
 	for (int i = 0; i < m_constraints.size(); ++i)
 	{
-			force.col(m_constraints[i]) = Eigen::Vector3d::Zero();
+			force.col(m_constraints[i]) = Eigen::Vector3f::Zero();
 			force.col(m_constraints[i]) += 1e8 * (restPos.col(m_constraints[i]) - pos.col(m_constraints[i]));
 	}
 
 	int n = pos.cols();
-	Eigen::SparseMatrix<double> A = (m_dumpingBelta + m_t)*m_t*K;
+	Eigen::SparseMatrix<float> A = (m_dumpingBelta + m_t)*m_t*K;
 	for (int i = 0; i < 3 * n; ++i)
 	{
 		A.coeffRef(i, i) += (1 + m_t*m_dumpingAlpha)*m_masses(i);
 	}
 
-	Eigen::Map<Eigen::VectorXd> p(pos.data(), n_nodes * 3);
-	Eigen::Map<Eigen::VectorXd> v(vel.data(), n_nodes * 3);
-	Eigen::Map<Eigen::VectorXd> f(force.data(), n_nodes * 3);
+	Eigen::Map<Eigen::VectorXf> p(pos.data(), n_nodes * 3);
+	Eigen::Map<Eigen::VectorXf> v(vel.data(), n_nodes * 3);
+	Eigen::Map<Eigen::VectorXf> f(force.data(), n_nodes * 3);
 
-	Eigen::VectorXd b = m_masses.cwiseProduct(v) + m_t * f;
+	Eigen::VectorXf b = m_masses.cwiseProduct(v) + m_t * f;
 
-	//Eigen::PardisoLDLT<Eigen::SparseMatrix<double>> pardiso_solver;
+	//Eigen::PardisoLDLT<Eigen::SparseMatrix<float>> pardiso_solver;
 	//omp_set_num_threads(8);
 	//mkl_set_num_threads(8);
 	mkl_set_dynamic(0);
@@ -184,14 +184,14 @@ void TimeIntegration::BackEuler(Eigen::MatrixXd & pos,
 	//std::cout << mkl_get_max_threads() << std::endl;
 }
 
-void TimeIntegration::addGroundConstraints(double y, Eigen::MatrixXd & pos, Eigen::MatrixXd & vel)
+void TimeIntegration::addGroundConstraints(float y, Eigen::MatrixXf & pos, Eigen::MatrixXf & vel)
 {
 	for (size_t i = 0; i < n_nodes; ++i)
 	{
 		if (pos(1, i) < y)
 		{
 			pos(1, i) = y;
-			vel.col(i) = Eigen::Vector3d::Zero();
+			vel.col(i) = Eigen::Vector3f::Zero();
 		}
 	}
 }
